Refactor kernel utilities
